Remove var from test (#8)
* Fix bug that creates extraneous gulp watchers

Each time gulp watch was run, it added 4 additional watchers to the
process. This creates a memory leak that eventually eats up all the
ram. The gulpfile and the gulp script in package.json was rewritten
to keep from creating unnecessary watchers.

* Remove var from repo

The original idea was to remove all reference of var from the test
but using let and const will break the test since it's re-run in
the same process. This lead to an Error being thrown where any
let or const variable has already been declared.

The fix was to run the foreclosure.js file in it's own context.
This lead to another bug where let and const variables could not
be pulled out of the newly created context. Thus an object is added
to the context for the student to place the three variables onto.
This does remove the need for var in the repo but ultimately does
not allow for the usage of let and const. But this could be used
as an opportunity to explain object references.